[Move] : Files Invoices.vue and SalesOrder.vue
